requirements.txt
fastapi
uvicorn
supabase
python-dotenv
bcrypt
duckduckgo_search
trafilatura
sentence-transformers
opencv-python-headless
razorpay
Pillow
imagehash
mediapipe
geopy
beautifulsoup4
httpx
torch
pandas
numpy
requests
tqdm
datasets
scikit-learn
onnxruntime
transformers
